Changing feeBPS from uint16 -> uint256 both in solidity (removing now the unnecessary cast to uint256), and in the corresponding tests. The 3 tests with explicit EIP712 signature values have had to be updated, as expected.

+++ contracts/Payments/IPaymentsERC20.sol
@@ -127,7 +127,7 @@ interface IPaymentsERC20 is IEIP712Verifier {
 
         // the percentage fee expressed in Basis Points (bps), typical in finance
         // Examples:  2.5% = 250 bps, 10% = 1000 bps, 100% = 10000 bps
-        uint16 feeBPS;
+        uint256 feeBPS;
 
         // the price of the asset, an integer expressed in the
         // lowest unit of the ERC20 token.
